requirements.txt
streamlit
opencv-python
numpy
pandas
vaderSentiment
deep-translator
Pillow
mediapipe
transformers
torch
scikit-learn
langdetect